get gbm tests to pass (but still doing the pointer cast trick to get uniqueness in the stubbuffers. should move those to a common file)

+++ tests/integration-tests/graphics/gbm/test_buffer_integration.cpp
@@ -42,7 +42,7 @@ class StubBuffer : public mc::BufferBasic
 {
 public:
     StubBuffer()
-     : BufferBasic(mc::BufferID{(int) this}),
+     : BufferBasic(mc::BufferID{(int) (long int) this}),
        creation_thread_id{std::this_thread::get_id()}
     {}
